# backend/ent/schema/playlistvideo.go
package schema

import (
	"github.com/facebook/ent"
	"github.com/facebook/ent/schema/edge"
)

// PlaylistVideo holds the schema definition for the PlaylistVideo entity.
type PlaylistVideo struct {
	ent.Schema
}

// Fields of the PlaylistVideo.
func (PlaylistVideo) Fields() []ent.Field {
	return []ent.Field{}
}

// Edges of the PlaylistVideo.
func (PlaylistVideo) Edges() []ent.Edge {
	return []ent.Edge{
		edge.From("video", Video.Type).Ref("playlist_videos").Unique(),
		edge.From("playlists", Playlist.Type).Ref("playlist_videos").Unique(),
		edge.From("resolution", Resolution.Type).Ref("playlist_videos").Unique(),
	}
}